requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.2
asn1crypto==1.5.1
beautifulsoup4==4.11.1
braintree==4.16.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
cssselect2==0.6.0
Django==4.0.6
django-bootstrap4==22.1
django-braintree==1.3.4
django-crispy-forms==1.14.0
django-mathfilters==1.0.0
django-phonenumber-field==6.3.0
django-shopping-cart==0.1
django-xhtml2pdf==0.0.4
future==0.18.2
html5lib==1.1
idna==3.3
lxml==4.9.1
MouseInfo==0.1.3
mysqlclient==2.1.1
oscrypto==1.3.0
phonenumbers==8.12.52
Pillow==9.2.0
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
PyMsgBox==1.0.9
PyPDF3==1.0.6
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-barcode==0.14.0
python-bidi==0.4.2
pytweening==1.0.4
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.11
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stripe==4.1.0
svglib==1.3.0
tinycss2==1.1.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.11
webencodings==0.5.1
xhtml2pdf==0.2.8
